final commit

+++ src/modules/auth/auth.service.ts
@@ -1,11 +1,12 @@
 import { Injectable, UnauthorizedException } from '@nestjs/common';
 import * as bcrypt from 'bcryptjs';
+import { Response, Request } from 'express';
 
 import { AuthCacheService } from './auth-cache.service';
 import { AuthResDto } from './dto/auth.res.dto';
 import { SignInReqDto } from './dto/sign-in.req.dto';
-import { TokenPairResDto } from './dto/token-pair.res.dto';
 import { TokenService } from './token.service';
+import { TokenType } from '../../database/enums/token-type.enum';
 import { IManagerData } from '../../interfaces/manager-data.interface';
 import { ManagerRepository } from '../repositories/services/manager.repository';
 import { RefreshTokenRepository } from '../repositories/services/refresh-token.repository';
@@ -19,7 +20,7 @@ export class AuthService {
     private readonly authCacheService: AuthCacheService,
   ) {}
 
-  public async signIn(dto: SignInReqDto): Promise<AuthResDto> {
+  public async signIn(dto: SignInReqDto, res: Response): Promise<AuthResDto> {
     const manager = await this.managerRepository.findOne({
       where: { email: dto.email },
       select: { id: true, password: true, role: true, surname: true },
@@ -55,15 +56,27 @@ export class AuthService {
       }),
       this.authCacheService.saveToken(tokens.accessToken, manager.id),
     ]);
+    this.setRefreshTokenCookie(res, tokens.refreshToken);
     manager.last_login = new Date();
     await this.managerRepository.save(manager);
     const managerEntity = await this.managerRepository.findOneBy({
       id: manager.id,
     });
-    return { manager: managerEntity, tokens };
+    return { manager: managerEntity, accessToken: tokens.accessToken };
   }
 
-  public async refresh(managerData: IManagerData): Promise<TokenPairResDto> {
+  public async refresh(
+    req: Request,
+    res: Response,
+  ): Promise<{ accessToken: string }> {
+    const refreshToken: string = req.cookies?.refreshToken;
+    if (!refreshToken) {
+      throw new UnauthorizedException('Refresh token is missing');
+    }
+    const managerData = await this.tokenService.verifyToken(
+      refreshToken,
+      TokenType.REFRESH,
+    );
     await Promise.all([
       this.refreshTokenRepository.delete({
         managerId: managerData.managerId,
@@ -88,15 +101,27 @@ export class AuthService {
         managerData.managerId,
       ),
     ]);
-
-    return tokens;
+    this.setRefreshTokenCookie(res, tokens.refreshToken);
+    return { accessToken: tokens.accessToken };
   }
-  public async signOut(managerData: IManagerData): Promise<void> {
+  public async signOut(
+    res: Response,
+    managerData: IManagerData,
+  ): Promise<void> {
     await Promise.all([
       this.refreshTokenRepository.delete({
         managerId: managerData.managerId,
       }),
       this.authCacheService.deleteToken(managerData.managerId),
     ]);
+    res.clearCookie('refreshToken');
+  }
+  private setRefreshTokenCookie(res: Response, refreshToken: string) {
+    res.cookie('refreshToken', refreshToken, {
+      httpOnly: true,
+      secure: true,
+      sameSite: 'strict',
+      maxAge: 7 * 24 * 60 * 60 * 1000,
+    });
   }
 }
